Postlist added

+++ src/main/java/com/example/individualwebservice/services/PostService.java
@@ -42,13 +42,17 @@ public class PostService {
 
         if (postInformation.getTopic() != null && postInformation.getContent() != null && id != 0) {
 
-            User existingUser = userService.findUserById(id);
+            User author = userService.findUserById(id);
 
-            if (existingUser != null) {
+            if (author != null) {
 
-                Post post = new Post(postInformation.getTopic(), postInformation.getContent(), existingUser);
+                Post post = new Post(postInformation.getTopic(), postInformation.getContent(), author);
+
+                author.addPostToList(post);
 
                 postRepository.save(post);
+
+                userService.saveOrUpdateUser(author);
 
                 return post;
             } else {
@@ -64,25 +68,56 @@ public class PostService {
     }
 
     public Post updatePost(long id, Post newPostInformation) {
-        Post existingPost = findPostById(id);
 
-        if (existingPost != null) {
+        if (newPostInformation.getAuthor() == null) {
 
-            if (!newPostInformation.getTopic().contains(existingPost.getTopic()) && !newPostInformation.getTopic().isEmpty()) {
+            Post existingPost = findPostById(id);
 
-                existingPost.setTopic(newPostInformation.getTopic());
-            }
-            if (!newPostInformation.getContent().contains(existingPost.getContent()) && !newPostInformation.getContent().isEmpty()) {
+            if (existingPost != null) {
 
-                existingPost.setContent(newPostInformation.getContent());
+                if (!newPostInformation.getTopic().equals(existingPost.getTopic()) && newPostInformation.getTopic() != null) {
+
+                    existingPost.setTopic(newPostInformation.getTopic());
+                }
+                if (!newPostInformation.getContent().equals(existingPost.getContent()) && newPostInformation.getContent() != null) {
+
+                    existingPost.setContent(newPostInformation.getContent());
+                }
             }
 
             postRepository.save(existingPost);
 
             return existingPost;
+
         } else {
 
-            return null;
+            Post existingPost = findPostById(id);
+            User existingAuthor = existingPost.getAuthor();
+            User newAuthor = userService.findUserById(newPostInformation.getAuthor().getId());
+
+            if (existingPost != null) {
+
+                if (!newPostInformation.getTopic().contains(existingPost.getTopic()) && newPostInformation.getTopic() != null) {
+
+                    existingPost.setTopic(newPostInformation.getTopic());
+                }
+                if (!newPostInformation.getContent().contains(existingPost.getContent()) && newPostInformation.getContent() != null) {
+
+                    existingPost.setContent(newPostInformation.getContent());
+                }
+
+                existingPost.setAuthor(newAuthor);
+
+                existingAuthor.removePostFromList(existingPost);
+                newAuthor.addPostToList(existingPost);
+
+                userService.saveOrUpdateUser(existingAuthor);
+                userService.saveOrUpdateUser(newAuthor);
+
+                postRepository.save(existingPost);
+            }
+
+            return existingPost;
         }
     }
 
@@ -91,9 +126,22 @@ public class PostService {
 
         if (postToBeDeleted != null) {
 
-            postRepository.delete(postToBeDeleted);
+            User author = postToBeDeleted.getAuthor();
 
-            return "Deleted post";
+            if (author != null) {
+
+                author.removePostFromList(postToBeDeleted);
+
+                userService.saveOrUpdateUser(author);
+
+                postRepository.delete(postToBeDeleted);
+
+                return "Deleted post";
+            } else {
+
+                return "ERROR: Issue when retrieving Author/User";
+            }
+
         } else {
 
             return "ERROR: Post ID provided does not exist";
